% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossModelComparisonUI.R
\name{crossModelComparisonUI}
\alias{crossModelComparisonUI}
\title{Cross Model Comparison UI Tab}
\usage{
crossModelComparisonUI()
}
\value{
A tabPanel for cross model comparison visualization
}
\description{
Cross Model Comparison UI Tab
}
\keyword{internal}
