% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureImportanceUI.R
\name{featureImportanceUI}
\alias{featureImportanceUI}
\title{Feature Importance UI Tab}
\usage{
featureImportanceUI()
}
\value{
A tabPanel for feature importance visualization
}
\description{
Feature Importance UI Tab
}
\keyword{internal}
