% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboard_launcher.R
\name{launch_amr_dashboard}
\alias{launch_amr_dashboard}
\title{Launch the AMR Dashboard}
\usage{
launch_amr_dashboard()
}
\description{
This function launches the AMR Shiny Dashboard from the 'inst/app/' directory
of the installed package.
}
