% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{launchAMRDashboard}
\alias{launchAMRDashboard}
\title{AMR Shiny dashboard application}
\usage{
launchAMRDashboard(results_root = NULL)
}
\arguments{
\item{results_root}{File path to the root directory containing amRml model
output results. If \code{NULL} (default), the application will attempt to
use example data bundled with the package, where available.}
}
\value{
A Shiny application object
}
\description{
Creates and returns a Shiny application for exploring antimicrobial
resistance data and machine learning model results.
}
\examples{
if (interactive()) {
  app <- launchAMRDashboard()
  shiny::runApp(app)
}
}
