% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelPerfUI.R
\name{modelPerfUI}
\alias{modelPerfUI}
\title{Model Performance UI Tab}
\usage{
modelPerfUI()
}
\value{
A tabPanel for model performance visualization
}
\description{
Model Performance UI Tab
}
\keyword{internal}
